+++ BPOR/BPOR.Domain/Entities/VolunteerStudyInformationGroupCriteria.cs
@@ -5,7 +5,7 @@ namespace BPOR.Domain.Entities;
 public class VolunteerStudyInformationGroupCriteria
 {
     public int Id { get; set; }
-    public long VolunteerStudyInformationGroupId { get; set; }
+    public int VolunteerStudyInformationGroupId { get; set; }
     public string Criteria { get; set; }
     
     public VolunteerStudyInformationGroup Group { get; set; }
